add updates from meeting
