+++ src/app/pages/home.jsx
@@ -1,0 +1,6 @@
+ import React from "react";
+ import ResponsiveAppBar from "../components/homeNavbar";
+
+
+ <ResponsiveAppBar/>
+
